% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem Function}
\usage{
birthday(n)
}
\arguments{
\item{n}{number of people}
}
\value{
probability that 2 or more people will have the same birthday given a certain number of people
}
\description{
Birthday Problem Function
}
\examples{
birthday(23) #gives probability that out of 23 people at least 2 will share a bday
}
