% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage in a suburb}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{x}{Distance the fire is from a fire station in miles}
  \item{y}{Damage caused by the fire in thousands of dollars}
  ...
}
}
\source{
Mendenhall and Sincich, STATISTICS for Engineering and the Sciences 6th Edition
}
\usage{
fire
}
\description{
Fire damage based on the distance a fire station is from the fire
}
\keyword{datasets}
