% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{sampling function}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
bar plot of probability of number of successes
}
\description{
sampling function
}
\examples{
mybin(n=15)
mybin(iter=500,n=10,p=0.5)
}
