% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2().R
\name{myboot2}
\alias{myboot2}
\title{bootstrapping function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{of bootstrap iterations (default=10000)}

\item{x}{numeric vector. Sample data to bootstrap}

\item{fun}{statistic to compute ('mean','var',etc.)}

\item{alpha}{significance level used to form confidence intervals}

\item{...}{other arguments passed to histogram function}
}
\value{
list with confidence interval, the statistic function, original data sampled, bootstrapped statistics over each iteration, and a graph with the statistics and confidence intervals on it.
}
\description{
bootstrapping function
}
\examples{
\dontrun{
set.seed(39)
sam <- rnorm(25, mean = 25, sd = 10)
myboot2(x = sam, fun = mean, alpha = 0.05)

# Using a custom statistic (mean divided by median)
ratio_fun <- function(x) mean(x) / median(x)
myboot2(x = sam, fun = ratio_fun, alpha = 0.30)
}
}
