% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Central Limit Theorem for the Binomial Distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{sample size for each iteration}

\item{iter}{# of iterations/how many samples to generate}

\item{p}{probability of a success}

\item{...}{whatever comes after}
}
\value{
a numeric vector containing sample means from each iteration, and it also returns a graphical representation of the distrubution of the sample means
}
\description{
Central Limit Theorem for the Binomial Distribution
}
\examples{
# Demonstrate CLT with Binomial distribution
set.seed(123)
mycltb(n = 10, iter = 10000, p = 0.3)
}
