% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{A function that returns the log-likelihood contribution of one observation}

\item{x}{A numeric vector containing the observed data.}

\item{param}{A numeric vector of parameter values over which the likelihood should be evaluated (the parameter grid).}

\item{...}{other parameters involving plots and whatnot}
}
\value{
A list with components for the MLE
}
\description{
Max Likelihood Function
}
\examples{
# Binomial example:
x <- c(3,3,4,3,4,5,5,4)
lfun <- function(x, p) dbinom(x, size = 20, prob = p, log = TRUE)
pgrid <- seq(0, 1, length = 500)
\dontrun{
  mymaxlik(lfun, x, pgrid, xlab = "p", ylab = "log-likelihood")
}

}
