% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Distribution Curve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{P(X<=a)}
}
\value{
A graph of a normal distribution curve with plotted region from
-inf. to a that represents the probability. Also returns mean, standard
deviation, and probability.
}
\description{
Normal Distribution Curve function
}
\examples{
myncurve(7,6,2) #returns a normal curve with mean 7, standard deviation 6,
#and has a shaded region from -inf. to 2 that represents the probability
#that X is less than 2
}
