% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{number of available seats}

\item{gamma}{probability of overbooking}

\item{p}{probability that a passenger shows}
}
\value{
the number of tickets to sell to make profit while avoiding overbooking, two plots to show the answer graphically, and the p, gamma, and N values (n values and plots have a discrete and continuous exmaple each)
}
\description{
Overbooking function
}
\examples{
ntickets(400,0.02,0.95)
}
